test(write): colocate fixed-bucket coverage
