+++ crates/paimon/src/table/postpone_fixed_bucket_write.rs
@@ -204,3 +204,366 @@ fn validate_postpone_fixed_bucket_write(table: &Table) -> Result<()> {
     }
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::catalog::Identifier;
+    use crate::spec::{DataType, IntType, Schema, TableSchema, VarCharType};
+    use crate::table::table_write::tests::{
+        make_batch, make_partitioned_batch_3col, read_id_value_rows, setup_dirs, test_file_io,
+        test_postpone_partitioned_table, test_postpone_pk_table,
+    };
+    use crate::table::{
+        CommitMessage, PostponeBucketPlan, SnapshotManager, Table, TableCommit, TableScan,
+    };
+    use arrow_array::{Int32Array, RecordBatch, StringArray};
+    use arrow_schema::{DataType as ArrowDataType, Field as ArrowField, Schema as ArrowSchema};
+    use std::sync::Arc;
+
+    fn make_partition_bucket_plan_batch(
+        partitions: Vec<&str>,
+        total_buckets: Vec<i32>,
+    ) -> RecordBatch {
+        RecordBatch::try_new(
+            Arc::new(ArrowSchema::new(vec![
+                ArrowField::new("pt", ArrowDataType::Utf8, false),
+                ArrowField::new("total_buckets", ArrowDataType::Int32, false),
+            ])),
+            vec![
+                Arc::new(StringArray::from(partitions)),
+                Arc::new(Int32Array::from(total_buckets)),
+            ],
+        )
+        .unwrap()
+    }
+
+    fn make_partition_bucket_plan(
+        table: &Table,
+        partitions: Vec<&str>,
+        total_buckets: Vec<i32>,
+    ) -> PostponeBucketPlan {
+        PostponeBucketPlan::from_arrow(
+            table,
+            &make_partition_bucket_plan_batch(partitions, total_buckets),
+        )
+        .unwrap()
+    }
+
+    fn make_unpartitioned_bucket_plan(table: &Table, total_buckets: i32) -> PostponeBucketPlan {
+        let batch = RecordBatch::try_new(
+            Arc::new(ArrowSchema::new(vec![ArrowField::new(
+                "total_buckets",
+                ArrowDataType::Int32,
+                false,
+            )])),
+            vec![Arc::new(Int32Array::from(vec![total_buckets]))],
+        )
+        .unwrap();
+        PostponeBucketPlan::from_arrow(table, &batch).unwrap()
+    }
+
+    async fn write_fixed_batch(
+        table: &Table,
+        commit_user: &str,
+        total_buckets: i32,
+        batch: &RecordBatch,
+    ) -> Vec<CommitMessage> {
+        let mut write = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user(commit_user)
+            .unwrap()
+            .with_bucket_plan(make_unpartitioned_bucket_plan(table, total_buckets))
+            .new_write()
+            .unwrap();
+        write.write_arrow_batch(batch).await.unwrap();
+        write.prepare_commit().await.unwrap()
+    }
+
+    fn assert_total_buckets(messages: &[CommitMessage], total_buckets: i32) {
+        assert!(!messages.is_empty());
+        assert!(messages
+            .iter()
+            .all(|message| message.total_buckets == Some(total_buckets)));
+    }
+
+    fn assert_one_shot_error(error: crate::Error) {
+        assert!(
+            matches!(error, crate::Error::DataInvalid { ref message, .. }
+            if message.contains("only supports one prepare_commit call")
+                && message.contains("create a new writer"))
+        );
+    }
+
+    #[tokio::test]
+    async fn test_postpone_batch_write_uses_visible_fixed_buckets() {
+        let file_io = test_file_io();
+        let table_path = "memory:/test_postpone_fixed_bucket_write";
+        setup_dirs(&file_io, table_path).await;
+        let table = test_postpone_pk_table(&file_io, table_path);
+
+        let first_builder = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user("fixed-user-1")
+            .unwrap()
+            .with_bucket_plan(make_unpartitioned_bucket_plan(&table, 2));
+        let mut write = first_builder.new_write().unwrap();
+        write
+            .write_arrow_batch(&make_batch(vec![1, 2, 3, 4], vec![10, 20, 30, 40]))
+            .await
+            .unwrap();
+        let messages = write.prepare_commit().await.unwrap();
+        assert!(messages.iter().all(|message| message.bucket >= 0));
+        assert_total_buckets(&messages, 2);
+
+        let stale_messages =
+            write_fixed_batch(&table, "stale-user", 1, &make_batch(vec![9], vec![90])).await;
+        assert_total_buckets(&stale_messages, 1);
+        TableCommit::new(table.clone(), "fixed-user-1".to_string())
+            .commit(messages)
+            .await
+            .unwrap();
+        let error = TableCommit::new(table.clone(), "stale-user".to_string())
+            .commit(stale_messages)
+            .await
+            .unwrap_err();
+        assert!(error.to_string().contains("Fixed-bucket conflict"));
+        assert_eq!(
+            read_id_value_rows(&table).await,
+            vec![(1, 10), (2, 20), (3, 30), (4, 40)]
+        );
+
+        let second_builder = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user("fixed-user-2")
+            .unwrap()
+            .with_bucket_plan(make_unpartitioned_bucket_plan(&table, 2));
+        let mut write = second_builder.new_write().unwrap();
+        write
+            .write_arrow_batch(&make_batch(vec![5], vec![50]))
+            .await
+            .unwrap();
+        let messages = write.prepare_commit().await.unwrap();
+        assert_total_buckets(&messages, 2);
+        assert_one_shot_error(
+            write
+                .write_arrow_batch(&make_batch(vec![6], vec![60]))
+                .await
+                .unwrap_err(),
+        );
+        assert_one_shot_error(write.prepare_commit().await.unwrap_err());
+        TableCommit::new(table.clone(), "fixed-user-2".to_string())
+            .commit(messages)
+            .await
+            .unwrap();
+        assert_eq!(
+            read_id_value_rows(&table).await,
+            vec![(1, 10), (2, 20), (3, 30), (4, 40), (5, 50)]
+        );
+    }
+
+    #[tokio::test]
+    async fn test_postpone_distributed_writers_share_bucket_plan() {
+        let file_io = test_file_io();
+        let table_path = "memory:/test_postpone_shared_bucket_plan";
+        setup_dirs(&file_io, table_path).await;
+        let table = test_postpone_partitioned_table(&file_io, table_path);
+        let plan = make_partition_bucket_plan(&table, vec!["p1", "p2"], vec![3, 3]);
+
+        let builder = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user("shared-plan-user")
+            .unwrap()
+            .with_bucket_plan(plan);
+        let mut first = builder.new_write().unwrap();
+        let mut second = builder.new_write().unwrap();
+        first
+            .write_arrow_batch(&make_partitioned_batch_3col(vec!["p1"], vec![1], vec![10]))
+            .await
+            .unwrap();
+        second
+            .write_arrow_batch(&make_partitioned_batch_3col(
+                vec!["p2", "p2", "p2", "p2"],
+                vec![2, 3, 4, 5],
+                vec![20, 30, 40, 50],
+            ))
+            .await
+            .unwrap();
+
+        let mut messages = first.prepare_commit().await.unwrap();
+        messages.extend(second.prepare_commit().await.unwrap());
+        assert_total_buckets(&messages, 3);
+        builder.new_commit().commit(messages).await.unwrap();
+    }
+
+    #[tokio::test]
+    async fn test_postpone_distributed_writers_reject_overlapping_bucket_ownership() {
+        let file_io = test_file_io();
+        let table_path = "memory:/test_postpone_overlapping_writer_ownership";
+        setup_dirs(&file_io, table_path).await;
+        let table = test_postpone_partitioned_table(&file_io, table_path);
+        let builder = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user("overlapping-writers")
+            .unwrap()
+            .with_bucket_plan(make_partition_bucket_plan(&table, vec!["p"], vec![1]));
+
+        let mut first = builder.new_write().unwrap();
+        let mut second = builder.new_write().unwrap();
+        first
+            .write_arrow_batch(&make_partitioned_batch_3col(vec!["p"], vec![1], vec![10]))
+            .await
+            .unwrap();
+        second
+            .write_arrow_batch(&make_partitioned_batch_3col(vec!["p"], vec![1], vec![20]))
+            .await
+            .unwrap();
+
+        let mut messages = first.prepare_commit().await.unwrap();
+        messages.extend(second.prepare_commit().await.unwrap());
+        let error = builder.new_commit().commit(messages).await.unwrap_err();
+        assert!(error
+            .to_string()
+            .contains("writer ownership conflict for bucket 0"));
+    }
+
+    #[tokio::test]
+    async fn test_postpone_provided_plan_must_cover_input_partitions() {
+        let file_io = test_file_io();
+        let table_path = "memory:/test_postpone_incomplete_bucket_plan";
+        setup_dirs(&file_io, table_path).await;
+        let table = test_postpone_partitioned_table(&file_io, table_path);
+
+        let error = PostponeBucketPlan::from_arrow(
+            &table,
+            &make_partition_bucket_plan_batch(vec!["p"], vec![0]),
+        )
+        .unwrap_err();
+        assert!(error.to_string().contains("must be positive"));
+
+        let plan = make_partition_bucket_plan(&table, vec!["p"], vec![2]);
+        let mut write = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_bucket_plan(plan)
+            .new_write()
+            .unwrap();
+
+        let error = write
+            .write_arrow_batch(&make_partitioned_batch_3col(
+                vec!["missing"],
+                vec![1],
+                vec![10],
+            ))
+            .await
+            .unwrap_err();
+        assert!(error
+            .to_string()
+            .contains("does not contain an input partition"));
+    }
+
+    #[tokio::test]
+    async fn test_postpone_overwrite_allows_bucket_rescale() {
+        let file_io = test_file_io();
+        let table_path = "memory:/test_postpone_overwrite_bucket_rescale";
+        setup_dirs(&file_io, table_path).await;
+        let table = test_postpone_partitioned_table(&file_io, table_path);
+
+        let initial_builder = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user("initial-layout")
+            .unwrap()
+            .with_bucket_plan(make_partition_bucket_plan(&table, vec!["p"], vec![1]));
+        let mut initial_write = initial_builder.new_write().unwrap();
+        initial_write
+            .write_arrow_batch(&make_partitioned_batch_3col(vec!["p"], vec![1], vec![10]))
+            .await
+            .unwrap();
+        initial_builder
+            .new_commit()
+            .commit(initial_write.prepare_commit().await.unwrap())
+            .await
+            .unwrap();
+
+        let overwrite_builder = table
+            .new_postpone_fixed_bucket_write_builder()
+            .unwrap()
+            .with_commit_user("replacement-layout")
+            .unwrap()
+            .with_bucket_plan(make_partition_bucket_plan(&table, vec!["p"], vec![3]))
+            .with_overwrite();
+        let mut overwrite_write = overwrite_builder.new_write().unwrap();
+        overwrite_write
+            .write_arrow_batch(&make_partitioned_batch_3col(vec!["p"], vec![2], vec![20]))
+            .await
+            .unwrap();
+        let messages = overwrite_write.prepare_commit().await.unwrap();
+        assert_total_buckets(&messages, 3);
+        overwrite_builder
+            .new_commit()
+            .commit(messages)
+            .await
+            .unwrap();
+
+        let snapshot = SnapshotManager::new(file_io, table_path.to_string())
+            .get_latest_snapshot()
+            .await
+            .unwrap()
+            .unwrap();
+        let entries = TableScan::new(&table, None, vec![], None, None, None)
+            .with_scan_all_files()
+            .plan_manifest_entries(&snapshot)
+            .await
+            .unwrap();
+        assert!(!entries.is_empty());
+        assert!(entries.iter().all(|entry| entry.total_buckets() == 3));
+    }
+
+    #[tokio::test]
+    async fn test_postpone_fixed_bucket_write_with_rowkind_field() {
+        let file_io = test_file_io();
+        let table_path = "memory:/test_postpone_fixed_bucket_rowkind";
+        setup_dirs(&file_io, table_path).await;
+        let schema = Schema::builder()
+            .column("id", DataType::Int(IntType::new()))
+            .column("value", DataType::Int(IntType::new()))
+            .column("op", DataType::VarChar(VarCharType::string_type()))
+            .primary_key(["id"])
+            .option("bucket", "-2")
+            .option("rowkind.field", "op")
+            .build()
+            .unwrap();
+        let table = Table::new(
+            file_io,
+            Identifier::new("default", "test_postpone_fixed_bucket_rowkind"),
+            table_path.to_string(),
+            TableSchema::new(0, &schema),
+            None,
+        );
+        let batch = RecordBatch::try_new(
+            Arc::new(ArrowSchema::new(vec![
+                ArrowField::new("id", ArrowDataType::Int32, false),
+                ArrowField::new("value", ArrowDataType::Int32, false),
+                ArrowField::new("op", ArrowDataType::Utf8, false),
+            ])),
+            vec![
+                Arc::new(Int32Array::from(vec![1])),
+                Arc::new(Int32Array::from(vec![10])),
+                Arc::new(StringArray::from(vec!["+I"])),
+            ],
+        )
+        .unwrap();
+
+        let messages = write_fixed_batch(&table, "fixed-rowkind", 1, &batch).await;
+
+        assert_eq!(messages.len(), 1);
+        assert_eq!(messages[0].bucket, 0);
+        assert_eq!(messages[0].total_buckets, Some(1));
+        assert_eq!(messages[0].new_files[0].row_count, 1);
+    }
+}
